Move the space discovery code to spaces
The previous space discovery code was at the wrong abstraction layer,
discovery is a property of a controller not in the persistence layer.

So it's natural that we move the old state code into a better location,
this code does that.
